cleanup wf buffer lib
